requirements.txt
# Core
pytz==2023.3
Django==4.2.1
django-configurations==2.4.1
gunicorn==20.1.0
newrelic==8.8.0
django-cors-headers==3.13.0

# For the persistence stores
psycopg2-binary==2.9.6
dj-database-url==2.0.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown
django-filter==23.2
django-formtools==2.5.1

# Developer Tools
ipdb==0.13.13
ipython
mkdocs==1.4.3
flake8==6.0.0

# Testing
mock==5.0.2
factory-boy==3.2.1
django-nose==1.4.7

coverage==7.2.5

# Static and Media Storage
django-storages==1.13.2
boto3==1.26.135

# django custum theme
djangoql==0.18.0


# Image processing
Pillow==9.0.0